+++ src/base.js
@@ -73,7 +73,7 @@ class WrhsCommand extends Command {
   /**
    * Builds a missing version error message
    *
-   * @returns {string} The missing host error message
+   * @returns {string} The missing version number error message
    */
   missingVersionError() {
     return 'Missing package version. Please sure package is in the form `packageName@version` where `version` is the specific version to build';